htm-monitor/requirements.txt
numpy
pytest
pyyaml
matplotlib